+++ src/Middleware/QueueHandler.php
@@ -11,7 +11,7 @@
  * @author Jörn Heinemann <joernheinemann@gmx.de>
  */
 
-namespace Gram\App;
+namespace Gram\Middleware;
 
 use Gram\Exceptions\MiddlewareNotAllowedException;
 use Psr\Container\ContainerInterface;
@@ -34,13 +34,13 @@ class QueueHandler implements RequestHandlerInterface
 
 	public function __construct(RequestHandlerInterface $last, ContainerInterface $container=null)
 	{
-		$this->last=$last;	//der rücksprung handler mit dem diese klasse aufgerufen wird
+		$this->last = $last;	//der rücksprung handler mit dem diese klasse aufgerufen wird
 		$this->container = $container;
 	}
 
 	public function add($middleware)
 	{
-		$this->stack[]=$middleware;	//nach jedem durchlauf wird ein element vom stack genommen
+		$this->stack[] = $middleware;	//nach jedem durchlauf wird ein element vom stack genommen
 	}
 
 	public function getLast()
@@ -67,14 +67,26 @@ class QueueHandler implements RequestHandlerInterface
 	{
 		//wenn es keine Middleware gibt gebe das Ergebnis des handlers aus der zuletzt getriggert werden soll
 
-		if(count($this->stack)===0){
+		if (count($this->stack) === 0) {
 			return $this->last->handle($request);
 		}
 
 		$middleware = \array_shift($this->stack);	//hole das oberste element und lösche es aus dem array
 
+		$middleware = $this->checkMiddleware($middleware);
+
+		return $middleware->process($request,$this);	//führe die middleware aus
+	}
+
+	/**
+	 * @param $middleware
+	 * @return MiddlewareInterface
+	 * @throws MiddlewareNotAllowedException
+	 */
+	protected function checkMiddleware($middleware)
+	{
 		//wenn ein Index für die Mw angegenen wurde, siehe im Container nach
-		if ($this->container!==null && \is_string($middleware)) {
+		if ($this->container !== null && \is_string($middleware)) {
 			if ($this->container->has($middleware) === false) {
 				throw new MiddlewareNotAllowedException("Middleware: [$middleware] not found");
 			}
@@ -86,6 +98,6 @@ class QueueHandler implements RequestHandlerInterface
 			throw new MiddlewareNotAllowedException("Middleware needs to implement Psr 15 MiddlewareInterface");
 		}
 
-		return $middleware->process($request,$this);	//führe die middleware aus
+		return $middleware;
 	}
 }